+++ src/smsfusion/__init__.py
@@ -3,13 +3,13 @@ from ._coning_sculling import ConingScullingAlg
 from ._ins import AHRS, VRU, AidedINS, FixedNED, StrapdownINS, gravity
 from ._smoothing import FixedIntervalSmoother
 from ._transforms import quaternion_from_euler
-from ._v2 import AHRSv2a
-from ._v2b import AHRSv2b
+from ._v2 import AHRSv2a, AHRSv2b, AHRSv2c
 
 __all__ = [
     "AHRS",
     "AHRSv2a",
     "AHRSv2b",
+    "AHRSv2c",
     "AidedINS",
     "benchmark",
     "constants",
