+++ src/smsfusion/_v2/_v2b.py
@@ -4,16 +4,16 @@ import numpy as np
 from numba import njit
 from numpy.typing import ArrayLike, NDArray
 
-from ._ins import _dhda_head, _h_head, _signed_smallest_angle
-from ._transforms import _angular_matrix_from_quaternion as T
-from ._transforms import _euler_from_quaternion, _rot_matrix_from_quaternion
+from .._ins import _dhda_head, _h_head, _signed_smallest_angle
+from .._transforms import _angular_matrix_from_quaternion as T
+from .._transforms import _euler_from_quaternion, _rot_matrix_from_quaternion
+from .._vectorops import _normalize, _skew_symmetric
 from ._v2common import (
     _correct_quat_with_gibbs2,
     _kalman_update_scalar,
     _kalman_update_sequential,
     _project_cov_ahead,
 )
-from ._vectorops import _normalize, _skew_symmetric
 
 VEL_IDX = slice(0, 3)
 ATT_IDX = slice(3, 6)
